# arquitetura-microsservicos/marketing-api/swagger_server/models/campanha_item.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CampanhaItem(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, id: str=None, name: str=None, public: str=None):  # noqa: E501
        """CampanhaItem - a model defined in Swagger

        :param id: The id of this CampanhaItem.  # noqa: E501
        :type id: str
        :param name: The name of this CampanhaItem.  # noqa: E501
        :type name: str
        :param public: The public of this CampanhaItem.  # noqa: E501
        :type public: str
        """
        self.swagger_types = {
            'id': str,
            'name': str,
            'public': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'public': 'public'
        }

        self._id = id
        self._name = name
        self._public = public

    @classmethod
    def from_dict(cls, dikt) -> 'CampanhaItem':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CampanhaItem of this CampanhaItem.  # noqa: E501
        :rtype: CampanhaItem
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this CampanhaItem.


        :return: The id of this CampanhaItem.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this CampanhaItem.


        :param id: The id of this CampanhaItem.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this CampanhaItem.


        :return: The name of this CampanhaItem.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this CampanhaItem.


        :param name: The name of this CampanhaItem.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def public(self) -> str:
        """Gets the public of this CampanhaItem.


        :return: The public of this CampanhaItem.
        :rtype: str
        """
        return self._public

    @public.setter
    def public(self, public: str):
        """Sets the public of this CampanhaItem.


        :param public: The public of this CampanhaItem.
        :type public: str
        """

        self._public = public